fix: Refactoring `MediaApi`
