+++ src/main/java/faang/school/postservice/api/media/MultipartFileMediaApi.java
@@ -1,8 +1,7 @@
-package faang.school.postservice.api;
+package faang.school.postservice.api.media;
 
 import faang.school.postservice.dto.media.MediaDto;
 import org.springframework.web.multipart.MultipartFile;
 
 public interface MultipartFileMediaApi extends MediaApi<String, MediaDto, MultipartFile> {
-    // TODO in future this interface was implemented for interact with S3
 }
