+++ src/main/java/faang/school/postservice/api/processor/MultipartFileProcessor.java
@@ -1,0 +1,6 @@
+package faang.school.postservice.api.processor;
+
+import org.springframework.web.multipart.MultipartFile;
+
+public interface MultipartFileProcessor extends ResourceProcessor<MultipartFile> {
+}
